Refactor: Moved JXL stuff into own package
