+++ src/commonMain/kotlin/com/ashampoo/kim/format/bmff/BoxReader.kt
@@ -18,17 +18,17 @@ package com.ashampoo.kim.format.bmff
 
 import com.ashampoo.kim.format.bmff.BMFFConstants.BMFF_BYTE_ORDER
 import com.ashampoo.kim.format.bmff.box.Box
-import com.ashampoo.kim.format.bmff.box.ExifBox
 import com.ashampoo.kim.format.bmff.box.FileTypeBox
 import com.ashampoo.kim.format.bmff.box.HandlerReferenceBox
 import com.ashampoo.kim.format.bmff.box.ItemInfoEntryBox
 import com.ashampoo.kim.format.bmff.box.ItemInformationBox
 import com.ashampoo.kim.format.bmff.box.ItemLocationBox
-import com.ashampoo.kim.format.bmff.box.JxlParticalCodestreamBox
 import com.ashampoo.kim.format.bmff.box.MediaDataBox
 import com.ashampoo.kim.format.bmff.box.MetaBox
 import com.ashampoo.kim.format.bmff.box.PrimaryItemBox
-import com.ashampoo.kim.format.bmff.box.XmlBox
+import com.ashampoo.kim.format.jxl.box.ExifBox
+import com.ashampoo.kim.format.jxl.box.JxlParticalCodestreamBox
+import com.ashampoo.kim.format.jxl.box.XmlBox
 import com.ashampoo.kim.input.PositionTrackingByteReader
 
 /**
