+++ src/commonMain/kotlin/com/ashampoo/kim/format/jxl/JxlReader.kt
@@ -1,27 +1,27 @@
 /*
- * Copyright 2024 Ashampoo GmbH & Co. KG
+ * Copyright 2023 Ashampoo GmbH & Co. KG
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
  */
-package com.ashampoo.kim.format.bmff
+package com.ashampoo.kim.format.jxl
 
 import com.ashampoo.kim.format.ImageMetadata
 import com.ashampoo.kim.format.bmff.box.Box
-import com.ashampoo.kim.format.bmff.box.ExifBox
-import com.ashampoo.kim.format.bmff.box.XmlBox
+import com.ashampoo.kim.format.jxl.box.ExifBox
+import com.ashampoo.kim.format.jxl.box.XmlBox
 import com.ashampoo.kim.model.ImageFormat
 
-internal object JxlHandler {
+internal object JxlReader {
 
     fun createMetadata(allBoxes: List<Box>): ImageMetadata {
 
@@ -29,8 +29,8 @@ internal object JxlHandler {
         val xmlBox = allBoxes.filterIsInstance<XmlBox>().firstOrNull()
 
         return ImageMetadata(
-            imageFormat = ImageFormat.JXL, // could be any ISO BMFF
-            imageSize = null, // not covered by ISO BMFF
+            imageFormat = ImageFormat.JXL,
+            imageSize = null, // TODO https://github.com/Ashampoo/kim/issues/65
             exif = exifBox?.tiffContents,
             exifBytes = exifBox?.exifBytes,
             iptc = null, // not covered by ISO BMFF
